Langchain-v1/pyproject.toml
[project]
name = "langchain-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.5.23",
    "google-generativeai>=0.8.4",
    "huggingface-hub>=0.27.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.15",
    "langchain-anthropic>=0.3.3",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.31",
    "langchain-google-genai>=2.0.9",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.14",
    "langchainhub>=0.1.21",
    "numpy>=1.26.4",
    "openai>=1.59.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "transformers>=4.46.3",
]
